serial/serialPort_forget-manual.https.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title></title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="resources/common.js"></script>
    <script src="resources/manual.js"></script>
  </head>
  <body>
    <p>
      These tests require a connected serial device.
    </p>
    <script>
      manual_serial_test(async (t, port) => {
        const portsBeforeForget = await navigator.serial.getPorts();
        assert_equals(portsBeforeForget.length, 1);
        assert_equals(portsBeforeForget[0], port);

        await port.forget();

        const portsAfterForget = await navigator.serial.getPorts();
        assert_equals(portsAfterForget.length, 0);
      }, 'forget() removes the device from getPorts()');
    </script>
  </body>
</html>
